pyproject.toml
[tool.poetry]
name = "misp-stix"
version = "2025.4.10"
description = "Python scripts used by MISP to export MISP format into STIX and to import STIX into MISP format."
authors = ["Christian Studer <christian.studer@circl.lu>"]
maintainers = ["Christian Studer <christian.studer@circl.lu>"]
license = "BSD-3-Clause"
repository = "https://github.com/MISP/misp-stix"
documentation = "https://misp.github.io/misp-stix/"
readme = "README.md"

classifiers = [
    'License :: OSI Approved :: BSD License',
    'Development Status :: 4 - Beta',
    'Environment :: Console',
    'Operating System :: POSIX :: Linux',
    'Intended Audience :: Science/Research',
    'Intended Audience :: Information Technology',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Programming Language :: Python :: 3.13',
    'Topic :: Security'
]

packages = [
    { include = "misp_stix_converter" }
]

include = [
    "misp_stix_converter/data/cti",
    "misp_stix_converter/data/misp-galaxy"
]

[tool.poetry.urls]
"Source" = "https://github.com/MISP/misp-stix"
"Bug Tracker" = "https://github.com/MISP/misp-stix/issues"

[tool.poetry.scripts]
misp_stix_converter = 'misp_stix_converter:main'

[tool.poetry.dependencies]
python = "^3.9"
pymisp = "^2.5.4"
stix = "^1.2.0"
cybox = "^2.1.0"
mixbox = "^1.0.3"
maec = "^4.1.0"
stix-edh = "1.0.3"
misp-lib-stix2 = "^3.0.1"
setuptools = "*"
stix2-validator = "^3.2.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^3.8.3"
pytest = "^7.1.2"

[build-system]
requires = ["poetry_core>=1.2", "setuptools"]
build-backend = "poetry.core.masonry.api"
